@@ -1,10 +1,10 @@
 package dev.nkucherenko.redischat.service;
 
 import dev.nkucherenko.redischat.dto.RoomDto;
-import dev.nkucherenko.redischat.entity.Room;
-
 import java.util.Set;
+import java.util.UUID;
 
 public interface RoomService {
     Iterable<RoomDto> getAllRoomsByIds(Set<String> roomIds);
+    RoomDto findById(UUID roomId);
 }
