@@ -6,14 +6,15 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.nkucherenko.redischat.dto.MessageDto;
 import dev.nkucherenko.redischat.dto.NotificationTypeEnum;
+import dev.nkucherenko.redischat.dto.RoomDto;
 import dev.nkucherenko.redischat.dto.UserNotificationDto;
 import dev.nkucherenko.redischat.entity.Message;
 import dev.nkucherenko.redischat.entity.Room;
 import dev.nkucherenko.redischat.entity.UserDetails;
 import dev.nkucherenko.redischat.mapper.MessageMapper;
-import dev.nkucherenko.redischat.repository.RoomRepository;
 import dev.nkucherenko.redischat.repository.UserRepository;
 import dev.nkucherenko.redischat.service.RedisListenerClient;
+import dev.nkucherenko.redischat.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.support.GenericApplicationContext;
@@ -30,7 +31,7 @@ import java.util.stream.Collectors;
 @Service
 public class RedisListenerClientImpl implements RedisListenerClient {
     private final GenericApplicationContext applicationContext;
-    private final RoomRepository roomRepository;
+    private final RoomService roomService;
     private final UserRepository userRepository;
     private final ObjectMapper objectMapper;
     private final MessageMapper messageMapper;
@@ -41,7 +42,7 @@ public class RedisListenerClientImpl implements RedisListenerClient {
         ChannelTopic channelTopic = resolveChannelTopic(resolvedChatId, applicationContext);
         UserDetails userDetails = userRepository.findById(UUID.fromString(userId)).orElseThrow();
         Set<ChannelTopic> channelTopicSet = userDetails.getRoomIds().stream().map(roomId -> {
-            Room room = roomRepository.findById(UUID.fromString(roomId)).orElse(new Room());
+            RoomDto room = roomService.findById(UUID.fromString(roomId));
             return resolveChannelTopic(room.getChatId(), applicationContext);
         }).collect(Collectors.toSet());
         channelTopicSet.add(channelTopic);
@@ -53,22 +54,15 @@ public class RedisListenerClientImpl implements RedisListenerClient {
         try {
             Message message = objectMapper.readValue(a.getMessage(), Message.class);
             MessageDto messageDto = messageMapper.messageToDto(message);
-
+            UserNotificationDto userNotificationDto = new UserNotificationDto()
+                    .setMessage(messageDto);
             if (message.getId() == null && message.getRoom() != null) {
-                log.info(message.toString());
-                return new UserNotificationDto()
-                        .setMessage(messageDto)
-                        .setType(NotificationTypeEnum.NEW_ROOM);
+                return userNotificationDto.setType(NotificationTypeEnum.NEW_ROOM);
             }
             if (resolvedChatId.equals(message.getChatId())) {
-                return new UserNotificationDto()
-                        .setMessage(messageDto)
-                        .setType(NotificationTypeEnum.CHAT_MESSAGE);
-            } else {
-                return new UserNotificationDto()
-                        .setMessage(messageDto)
-                        .setType(NotificationTypeEnum.NOTIFICATION_MESSAGE);
+                return userNotificationDto.setType(NotificationTypeEnum.CHAT_MESSAGE);
             }
+            return userNotificationDto.setType(NotificationTypeEnum.NOTIFICATION_MESSAGE);
         } catch (JsonProcessingException jsonProcessingException) {
             return new UserNotificationDto();
         }
